Revert "Fix: Trigger getInitialProps from client-side"
