+++ pages/Homepage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import App from '../components/layout/App';
+
+const Homepage = () => (
+  <div>
+    <App />
+  </div>
+);
+
+export default Homepage;
